feat: added deleting old message logic
- getting redis from state
- getting old message id from redis
- put old message id to redis

+++ bot/tasks/post_rate.py
@@ -6,6 +6,7 @@ from aiogram.exceptions import TelegramBadRequest
 from core import message_texts
 from core.config import settings
 from core.taskiq import broker
+from redis.asyncio import Redis
 from taskiq import Context, TaskiqDepends
 from utils.coinex_api import CoinexAPI
 
@@ -18,6 +19,7 @@ async def publish_usdt_rub_price(
     bot: Bot = TaskiqDepends(),
 ) -> None:
     coinex_api: CoinexAPI = context.state.coinex_api
+    redis: Redis = context.state.redis
 
     response = await coinex_api.get_usdt_rub_price()
 
@@ -26,10 +28,15 @@ async def publish_usdt_rub_price(
         text=message_texts.PRICE.format(price=response.ask_rate),
         disable_notification=True,
     )
-    try:
-        await bot.delete_message(
-            chat_id=settings.channel_id,
-            message_id=message.message_id - 1,
-        )
-    except TelegramBadRequest as ex:
-        logger.error(ex, exc_info=ex)
+
+    old_message_id = await redis.get("old_message_id")
+    if old_message_id:
+        try:
+            await bot.delete_message(
+                chat_id=settings.channel_id,
+                message_id=old_message_id,
+            )
+        except TelegramBadRequest as ex:
+            logger.error(msg="Error while deleting old msg", exc_info=ex)
+
+    await redis.set(name="old_message_id", value=message.message_id)
